move bigintToBigNumber to a separate file
