+++ packages/cli/src/packages-to-be/governance.ts
@@ -4,9 +4,8 @@ import { Address, bufferToHex, zip } from '@celo/base'
 import BigNumber from 'bignumber.js'
 import { gte } from 'semver'
 import { PublicClient } from 'viem'
-import { bigintToBigNumber } from '../utils/checks'
 import { resolveAddress } from './address-resolver'
-import { isCel2 } from './utils'
+import { bigintToBigNumber, isCel2 } from './utils'
 
 export interface ProposalMetadata {
   proposer: Address
